pyproject.toml
# pyproject.toml
[build-system]
requires      = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "dbt-schema-gen"
version         = "0.1.0"
description     = "Generate dbt schema.yml files with an LLM"
authors         = [{ name = "hdser"}]
readme          = "README.md"
license         = { file = "LICENSE" }
requires-python = ">=3.9"

dependencies = [
  "click>=8.1.7",
  "openai>=1.14.2",
  "anthropic>=0.25.3",
  "google-generativeai>=0.5.0",
  "pyyaml>=6.0",
  "sqlparse>=0.4.4",
  "python-dotenv>=1.0.1",
  "jinja2>=3.1.2",
]

[project.optional-dependencies]
dev = ["pytest", "black", "ruff"]

[project.scripts]
dbt-schema-gen = "dbt_schema_gen.cli:cli"
